+++ website/server/package.json
@@ -7,7 +7,8 @@
     "build": "tsc",
     "bundle": "node scripts/bundle.mjs",
     "lint": "npm run lint-tsc",
-    "lint-tsc": "tsgo --noEmit",
+    "lint-tsc": "tsgo --noEmit -p tsconfig.test.json",
+    "test": "vitest",
     "start": "node dist/index.js",
     "start:bundle": "node dist-bundled/server.mjs",
     "clean": "rimraf dist dist-bundled",
@@ -31,6 +32,7 @@
     "rimraf": "^6.1.3",
     "tsx": "^4.21.0",
     "typescript": "^5.9.3",
+    "vitest": "^4.1.4",
     "rolldown": "^1.0.0-rc.15",
     "rollup-plugin-swc3": "^0.12.1"
   }
